fix: resolve 185 bugs from comprehensive correctness audit (round 16)
8-agent parallel audit across all stack layers found and fixed 185 bugs
comparing against reference implementations, libsrtp2, and RFCs
3711/3550/3261/4733/5389/5766/3389/4566/3611/4028/2617/5761/4961/2833/6337.

P0 critical (13 fixed):
- SRTP: key zeroization, replay window reset attack, AES-CM counter overflow,
  timing side-channel in constant_time_eq
- RTP engine: symmetric RTP integration, SRTP send/recv wiring, RTCP-mux
  demux fix, ROC tracking, RTCP session integration
- Dead code: goertzel/vad/media_tap/plc module declarations added
- Jitter: BTreeMap replaced with HashMap, div-by-zero guard
- DNS: removed fake SRV queries, fixed SIPS port/transport handling

P1 high (83 fixed):
- SRTP: non-monotonic send rejection, 128-bit replay window, key lifetime
  enforcement, SSRC change detection, generic error messages
- RTCP: MAX_DROPOUT/MAX_MISORDER validation, padding handling, XR parsing,
  RTT monotonic clock, jitter buffer metrics
- SIP: per-provider registration, RTP/SAVP+crypto in SDP, Via transport,
  even port allocation, port-in-use tracking, Require header/420 response,
  re-INVITE codec renegotiation, incoming REFER, session timer Min-SE/422
- DTMF: wraparound threshold, END packet spacing, A-law silence detection,
  IGNORE_DURATION and CISCO_SKIP_MARKER flags
- Jitter: duplicate detection, timestamp jump reset, NACK filtering,
  adaptive delay hysteresis, playout timer for PLC
- NAT: STUN FINGERPRINT+MESSAGE-INTEGRITY, source validation, shared socket
  buffer, Symmetric NAT->Relay, TURN 438 nonce refresh, IPv4/IPv6 pool
- Auth: MD5-sess, username escaping, realm validation, nonce TTL
- State: unified CallState enum, validator wired in

P2 minor (89 fixed):
- All remaining items across SRTP, RTCP, SIP, DTMF, jitter, PLC, NAT,
  config, session management, provider routing

Also removes old gap analysis files and vendor-specific references.
550 tests passing, 0 failures (up from 212 tests previously).

+++ src/rtp/vad.rs
@@ -350,8 +350,11 @@ impl VoiceActivityDetector {
                     })
                     .sum::<i64>()
                     / per_ch as i64;
-                // Bug #25: Use f64 intermediate for division to avoid precision loss
-                self.dc_estimate = ((self.dc_estimate as f64 * 255.0) / 256.0) as i64 + frame_avg;
+                // P2-DTMF-6: Use purely integer arithmetic for DC estimate EMA.
+                // dc_estimate is scaled by 256. The EMA formula is:
+                //   dc = dc * 255 / 256 + frame_avg
+                // Using integer division with rounding toward zero.
+                self.dc_estimate = (self.dc_estimate * 255) / 256 + frame_avg;
             }
         }
